requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
PyJWT==2.8.0
werkzeug==2.3.7 